Changes to the code in order for stats to be kept.
