+++ software/src/main.cpp
@@ -1,11 +1,18 @@
 #include <iostream>
 #include <fstream>
+#include <thread>
 
 #include "common.hpp"
 #include "species.hpp"
 #include "network.hpp"
+#include "node.hpp"
+#include "gene.hpp"
 
 using namespace std;
+
+mutex stats_mutex;
+
+void monitorAI(list<Species *>);
 
 int main(int argc, char **argv, char **env){
    cout << "[INFO][MAIN]:\t Entered main." << endl;
@@ -15,6 +22,7 @@ int main(int argc, char **argv, char **env){
    exoAI.front()->add_network(new Network());
 
    while(true){   
+
       for(list<Species *>::iterator it = exoAI.begin(); it != exoAI.end(); ++it){
          (*it)->run_networks();
 
@@ -29,7 +37,7 @@ int main(int argc, char **argv, char **env){
          }
       }
    }
-
+   
    cout << "[INFO][MAIN]:\t Exiting main." << endl;
    return 0;
 }
